Refactor Streamlit app into modules
