+++ src/pixelcat/__init__.py
@@ -1,0 +1,2 @@
+"""Pixelcat analysis helpers."""
+
