+++ src/pixelcat/ui/__init__.py
@@ -1,0 +1,1 @@
+"""Streamlit UI helpers for Pixelcat."""
